Remove manual save (#13)
* Remove manual save

Now calling .set() will just automatically save the current data to the
persister. This removes the concept of an "unsaved" fact graph, which
causes a lot of footguns in the interface.

Eventually I would like to to remove the concept of a persister
entirely, which would dramaticaly simplify the fact graph. For now
though, just calling .save() on set will suffice.

* Surface limit violations on set